NOTE: THIS WAS CREATE BY JORDAN AND KYLE, NOT JUST JORDAN
Set up the opponents screen as a state within the main board screen, it shows the opponents active card and  has a button which will show further details of the opponent card. Working on creating backs of cards so you will also see face sown representaions of cards for the bench and deck.
Minor problem with showing opponents button and navigating to the state (no errors but not fully functional)