Done by Kyle Bell (40158884) and Jordan McComb (40156883)
Changed the menu screen background image and play button image so that it looks more appealing. Last menu screen looked stretched.

Changed the position of the play button on the screen so that it better fit the background.

Note: Throughout the last few weeks Jordan and I have also been working on improving the art and graphics of the game. Lots of card artwork has been submitted to the team to be implemented in the JSON, although time may not permit these being added.