Done by Kyle Bell (40158884) and Jordan McComb (40156883)
Changed the bitmaps for the grid tiles to make the game more lively and they now fit the genre/theme of the game. Also changed the opponent button, graveyard and deck bitmaps in the playarea to make they fit the theme of the game as well. All bitmap changes have now been submitted.