Done by Jordan McComb (40156883) and Kyle Bell (40158884)
This folder contains developmental assets used throughout the project.
All cards seen here were made by Jordan and Kyle.
Not all cards have been used but were considered throughout the projects.
The 3rd move (or in some casews fourth) for cards describe status effect moves, which we were ended up deciding not to implement in the end. (Due to complexity and time constraints)

This folder also contains the rulebook. The team composed this before we started the project so that we had a clear idea from the satrt how the game was to be played.
This rulebook details how to play the game and also shows some early concept art, some of which ended up in the final project